requirements.txt
firebase-admin~=6.5.0
pandas~=2.2.3
requests~=2.32.3
streamlit
python-dotenv